requirements.txt
fastapi
uvicorn
websockets
python-multipart
aiofiles
aiohttp
async-timeout
attrs
click
httpx
Jinja2
langchain
langchain-core
openai
orjson
pydantic
python-dotenv
PyYAML
supabase
SQLAlchemy
tqdm
langchain-community
langchain-openai
